@@ -1,6 +1,6 @@
 {
   "name": "cityvizor",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "description": "",
   "homepage": "https://cityvizor.cz/",
   "repository": "otevrena-data-mfcr/CityVizor",
@@ -10,7 +10,7 @@
     "build": "ng build --prod --app=0 && ng build --prod --app=1",
     "prod": "NODE_ENV=production node server/server.js",
     "cron": "NODE_ENV=production node server/cron.js",
-    "dev": "NODE_ENV=development concurrently \"ng build --watch --dev --progress false\" \"nodemon server/server.js\""
+    "dev": "NODE_ENV=development concurrently \"ng build --watch --aot --progress false\" \"nodemon server/server.js\""
   },
   "keywords": [
     "opendata",
@@ -18,64 +18,64 @@
     "faktury"
   ],
   "dependencies": {
-    "@angular/animations": "^4.3.6",
-    "@angular/common": "^4.3.6",
-    "@angular/compiler": "^4.3.6",
-    "@angular/core": "^4.3.6",
-    "@angular/forms": "^4.3.6",
-    "@angular/http": "^4.3.6",
-    "@angular/platform-browser": "^4.3.6",
-    "@angular/platform-browser-dynamic": "^4.3.6",
-    "@angular/router": "^4.3.6",
-    "angular2-jwt": "^0.2.3",
-    "archiver": "^1.3.0",
+    "@angular/animations": "^5.0.1",
+    "@angular/common": "^5.0.1",
+    "@angular/compiler": "^5.0.1",
+    "@angular/core": "^5.0.1",
+    "@angular/forms": "^5.0.1",
+    "@angular/http": "^5.0.1",
+    "@angular/platform-browser": "^5.0.1",
+    "@angular/platform-browser-dynamic": "^5.0.1",
+    "@angular/platform-server": "^5.0.1",
+    "@angular/router": "^5.0.1",
+    "@auth0/angular-jwt": "^1.0.0-beta.9",
+    "archiver": "^2.1.0",
     "batch-stream": "^0.1.3",
     "bcrypt": "^1.0.3",
-    "body-parser": "^1.17.0",
+    "body-parser": "^1.18.2",
     "bootstrap": "^3.3.6",
     "cheerio": "^1.0.0-rc.2",
     "classlist.js": "^1.1.20150312",
-    "compression": "^1.7.0",
-    "core-js": "^2.5.0",
-    "cron": "^1.2.1",
-    "csv-parse": "^1.2.1",
-    "express": "^4.15.4",
-    "express-dynacl": "github:smallhillcz/express-dynacl",
+    "compression": "^1.7.1",
+    "core-js": "^2.5.1",
+    "cron": "^1.3.0",
+    "csv-parse": "^2.0.0",
+    "express": "^4.16.2",
+    "express-dynacl": "^1.0.0",
     "express-jsonschema": "^1.1.6",
     "express-jwt": "^5.1.0",
     "https": "^1.0.0",
     "intl": "^1.2.5",
-    "jsonwebtoken": "^7.4.3",
+    "jsonwebtoken": "^8.1.0",
     "minimist": "^1.2.0",
-    "moment": "^2.18.1",
-    "mongodb": "^2.2.31",
-    "mongoose": "^4.11.9",
+    "moment": "^2.19.2",
+    "mongodb": "^2.2.33",
+    "mongoose": "^4.13.2",
     "mongoose-paginate": "^5.0.3",
     "mongoose-write-stream": "0.0.0",
     "multer": "^1.3.0",
     "ng2-file-upload": "^1.2.1",
-    "ngx-bootstrap": "^1.9.1",
+    "ngx-bootstrap": "^2.0.0-beta.8",
     "path": "^0.12.7",
     "reflect-metadata": "^0.1.10",
-    "rxjs": "^5.4.3",
+    "rxjs": "^5.5.2",
     "simple-get": "^2.7.0",
-    "stream-transform": "^0.1.2",
+    "stream-transform": "^1.0.0",
     "web-animations-js": "^2.3.1",
     "xml2js": "^0.4.19",
-    "zone.js": "^0.8.17"
+    "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "^1.3.2",
-    "@angular/compiler-cli": "^4.3.6",
-    "@angular/language-service": "^4.3.6",
-    "@types/node": "^6.0.88",
+    "@angular/cli": "^1.5.0",
+    "@angular/compiler-cli": "^5.0.1",
+    "@types/node": "^8.0.51",
     "codelyzer": "~3.0.1",
     "concurrently": "^3.0.0",
-    "mongo-express": "^0.34.0",
-    "nodemon": "^1.11.0",
+    "mongo-express": "^0.44.0",
+    "nodemon": "^1.12.1",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "typescript": "^2.4.2"
   },
   "author": "Martin Kopecek <martin.kopecek@smallhill.cz>",
   "license": "GPL-3.0"
