@@ -1,9 +1,4 @@
 module.exports = {
-	"*": true,
-	"profiles": {
-		"read": req => {
-			req.query.hidden = true;
-			return true;
-		}
-	}
+	// access all
+	"*": true
 };
