@@ -1,3 +1,3 @@
 
 // backwards compatibility issue, TODO: check all files and replace
-export { AppConfig } from "./config";
+export * from "./config";
