@@ -1,25 +1,45 @@
 
-var AppConfig = {
+// config injection token, i. e. the key in the injection storage map
+import { InjectionToken } from '@angular/core';
+export interface IAppConfig { [k:string]:any; }
+export let AppConfig = new InjectionToken<IAppConfig>('app.config');
+
+
+
+/* CONFIGURATION */
+
+export interface Module {
+	id: string;
+	url: string;
+	name: string;
+	optional: boolean;
+}
+
+let modules:Module[] = [
+  {"id": "expenditure-viz", "url": "vydaje", "name": "Výdaje", "optional": true},
+  {"id": "income-viz", "url": "prijmy", "name": "Příjmy", "optional": true},	
+  {"id": "invoice-list", "url": "faktury", "name": "Faktury", "optional": true},	
+  {"id": "notice-board", "url": "uredni-deska", "name": "Úřední deska", "optional": true},
+  {"id": "contract-list", "url": "registr-smluv", "name": "Registr smluv", "optional": true},
+  {"id": "data-sources", "url": "data", "name": "Data", "optional": true},
+];
+
+// configuration data
+export const AppConfigData: IAppConfig = {
 
   title: "CityVizor.cz",
   // root directory for static documents referenced in various links throughout the site
   docsUrl: "https://otevrena-data-mfcr.github.io/CityVizor",
 
   avatarsUrl: "/data/uploads/avatars",
+  
+  GitHub: {url: "https://github.com/otevrena-data-mfcr/CityVizor"},
 
   // infomail used in texts
   mail: "cityvizor@otevrenamesta.cz",
 
   // modules and their names. this controls what shows in menus etc.
-  modules: [
-    {"id": "dash-board", "url": "prehled", "name": "Aktuálně", "optional": false},
-    {"id": "expenditure-viz", "url": "vydaje", "name": "Výdaje", "optional": true},
-    {"id": "income-viz", "url": "prijmy", "name": "Příjmy", "optional": true},	
-    {"id": "invoice-list", "url": "faktury", "name": "Faktury", "optional": true},	
-    {"id": "notice-board", "url": "uredni-deska", "name": "Úřední deska", "optional": true},
-    {"id": "contract-list", "url": "registr-smluv", "name": "Registr smluv", "optional": true},
-    {"id": "data-sources", "url": "data", "name": "Data", "optional": true},
-  ],
+  modules: modules,
 
   acl: {
     
@@ -57,5 +77,3 @@ var AppConfig = {
   }
 
 }
-
-export { AppConfig };
