@@ -1,11 +1,6 @@
 import { Injectable } from '@angular/core';
-import { Http, URLSearchParams } from '@angular/http';
-import { AuthHttp } from 'angular2-jwt';
+import { HttpClient } from '@angular/common/http';
 import 'rxjs/add/operator/toPromise';
-
-import { FileUploader } from 'ng2-file-upload/file-upload/file-uploader.class';
-
-import { AuthService } from "./auth.service";
 
 function toParams(options){
 	if(!options) return "";
@@ -27,88 +22,87 @@ function toParams(options){
 	* saveEntity - updates entity and returns Promise with updated data on entity
 	* getDashboard - returns Promise with dashboard data
 	* getExpenditures - returns Promise with expenditures data for entity and year
-	* getExpendituresUploader - returns ng2-file-uploader FileUploader class ready to be attached to input or dropzone
 	**/
 @Injectable()
 export class DataService {
 	
-	constructor(private http: AuthHttp, private authService:AuthService) { }
+	constructor(private http: HttpClient) { }
 	
 	
 	/* PROFILES */
 	getProfiles(options?) {
-			return this.http.get("/api/profiles" + toParams(options)).toPromise().then(response => response.json());
+			return this.http.get<any>("/api/profiles" + toParams(options)).toPromise();
 	}
 
 	getProfile(profileId) {
-		return this.http.get("/api/profiles/" + profileId).toPromise().then(response => response.json());
+		return this.http.get<any>("/api/profiles/" + profileId).toPromise();
 	}
 	
 	createProfile(profile){
-		return this.http.post("/api/profiles",profile).toPromise().then(response => response.json());
+		return this.http.post<any>("/api/profiles",profile).toPromise();
 	}
 	
 	saveProfile(profile){
-		return this.http.put("/api/profiles/" + profile._id,profile).toPromise().then(response => response.json());
+		return this.http.put<any>("/api/profiles/" + profile._id,profile).toPromise();
 	}
 	
 	saveProfileAvatar(profileId,data:FormData){
 		return this.http.put("/api/profiles/" + profileId + "/avatar",data).toPromise();
 	}
 	deleteProfileAvatar(profileId){
-		return this.http.delete("/api/profiles/" + profileId + "/avatar").toPromise();
+		return this.http.delete("/api/profiles/" + profileId + "/avatar", { responseType: 'text' }).toPromise();
 	}
 	
 	getProfileBudget(profileId,year){
-		return this.http.get("/api/profiles/" + profileId + "/budgets/" + year).toPromise().then(response => response.json());
+		return this.http.get<any>("/api/profiles/" + profileId + "/budgets/" + year).toPromise();
 	}
 	saveProfileBudget(profileId,year,data:FormData){
-		return this.http.put("/api/profiles/" + profileId + "/budgets/" + year,data).toPromise().then(response => response.json());
+		return this.http.put<any>("/api/profiles/" + profileId + "/budgets/" + year,data).toPromise();
 	}
 	deleteProfileBudget(profileId,year){
-		return this.http.delete("/api/profiles/" + profileId + "/budgets/" + year).toPromise();
+		return this.http.delete("/api/profiles/" + profileId + "/budgets/" + year, { responseType: 'text' }).toPromise();
 	}
 	
 	getProfileBudgets(profileId,options?){
-		return this.http.get("/api/profiles/" + profileId + "/budgets" + toParams(options)).toPromise().then(response => response.json());
+		return this.http.get<any[]>("/api/profiles/" + profileId + "/budgets" + toParams(options)).toPromise();
 	}
 	
 	getProfileContracts(profileId,options?){
-		return this.http.get("/api/profiles/" + profileId + "/contracts" + toParams(options)).toPromise().then(response => response.json());
+		return this.http.get<any>("/api/profiles/" + profileId + "/contracts" + toParams(options)).toPromise();
 	}
 	
 	getProfileDashboardDashboard(profileId){
-		return this.http.get("/api/profiles/" + profileId + "/dashboard").toPromise().then(response => response.json());
+		return this.http.get<any>("/api/profiles/" + profileId + "/dashboard").toPromise();
 	}
 
 	getProfileEvents(profileId,options?){
-		return this.http.get("/api/profiles/" + profileId + "/events" + toParams(options)).toPromise().then(response => response.json());
+		return this.http.get<any[]>("/api/profiles/" + profileId + "/events" + toParams(options)).toPromise();
 	}
 	
 	getProfilePayments(profileId,options?){
-		return this.http.get("/api/profiles/" + profileId + "/payments" + toParams(options)).toPromise().then(response => response.json());
+		return this.http.get<any>("/api/profiles/" + profileId + "/payments" + toParams(options)).toPromise();
 	}
 	getProfilePaymentsMonths(profileId){
-		return this.http.get("/api/profiles/" + profileId + "/payments/months").toPromise().then(response => response.json());
+		return this.http.get<any[]>("/api/profiles/" + profileId + "/payments/months").toPromise();
 	}
 	
 	getProfileManagers(profileId){
-		return this.http.get("/api/profiles/" + profileId + "/managers").toPromise().then(response => response.json());
+		return this.http.get<any[]>("/api/profiles/" + profileId + "/managers").toPromise();
 	}
 	
 	getProfileNoticeBoard(profileId,options?){
-		return this.http.get("/api/profiles/" + profileId + "/noticeboard" + toParams(options)).toPromise().then(response => response.json());
+		return this.http.get<any[]>("/api/profiles/" + profileId + "/noticeboard" + toParams(options)).toPromise();
 	}
 	
 
 
 	/* EVENTS */
 	getEvent(eventId){
-		return this.http.get("/api/events/" + eventId).toPromise().then(response => response.json());
+		return this.http.get<any>("/api/events/" + eventId).toPromise();
 	}
 	
 	getEvents(options?){
-		return this.http.get("/api/events" + toParams(options)).toPromise().then(response => response.json());
+		return this.http.get<any[]>("/api/events" + toParams(options)).toPromise();
 	}
 	
 	
@@ -118,15 +112,15 @@ export class DataService {
 	
 	/* USERS */	
 	getUsers(){
-		return this.http.get("/api/users").toPromise().then(response => response.json());
+		return this.http.get<any>("/api/users").toPromise();
 	}
 	
 	getUser(userId){
-		return this.http.get("/api/users/" + userId).toPromise().then(response => response.json());
+		return this.http.get<any>("/api/users/" + userId).toPromise();
 	}
 	
 	saveUser(userData){
-		return this.http.post("/api/users/" + userData._id,userData).toPromise().then(response => response.json());
+		return this.http.post<any>("/api/users/" + userData._id,userData).toPromise();
 	}
 	
 	deleteUser(userId){
@@ -135,11 +129,11 @@ export class DataService {
 	
 	/* ETLs */
 	getETLs(options?){
-		return this.http.get("/api/etl" + toParams(options)).toPromise().then(response => response.json());	
+		return this.http.get<any>("/api/etl" + toParams(options)).toPromise();	
 	}
 	
 	getLatestETLs(profileId,options?){
-		return this.http.get("/api/etl/latest/" + profileId + toParams(options)).toPromise().then(response => response.json());	
+		return this.http.get<any>("/api/etl/latest/" + profileId + toParams(options)).toPromise();	
 	}
 	
 }
