@@ -5,9 +5,10 @@ import { Subscription } from 'rxjs/Subscription' ;
 import { ModalDirective } from 'ngx-bootstrap';
 
 import { DataService } from '../../../services/data.service';
+import { CodelistService } from "../../../services/codelist.service";
 import { ToastService } 		from '../../../services/toast.service';
 
-import { itemNames } from '../../../shared/data/item-names.data';
+import { ItemNamesCodelist } from '../../../shared/schema/codelist';
 
 import { ChartGroups }  from "../../../shared/data/chartGroups";
 
@@ -19,51 +20,51 @@ Component for graphical vizualization of expenditures
 @Component({
 	moduleId: module.id,
 	selector: 'income-viz',
-  host: {'(window:keydown)': 'hotkeys($event)'},
-	templateUrl: 'income-viz.template.html',
-	styleUrls: ['../../../shared/styles/inc-exp-viz.style.css']
-})
+	host: {'(window:keydown)': 'hotkeys($event)'},
+					 templateUrl: 'income-viz.template.html',
+					 styleUrls: ['../../../shared/styles/inc-exp-viz.style.css']
+					 })
 export class IncomeVizComponent{
-	
+
 	/* DATA */
 	@Input()
 	profile:any;
-	
+
 	profileId: string;
-	
+
 	@ViewChild('eventReceiptsModal')
 	public eventReceiptsModal:ModalDirective;
 	modalLoaded = false;
 
 	state:{year:number,group:string,event:string} = {year:null,group:null,event:null};
-	
+
 	groups: any[] = [];
 	groupIndex: any = {};
 
 	events:any[];
 	eventIndex:any = {};
-	itemNames = itemNames;
+	itemNames:ItemNamesCodelist;
 
 	maxAmount:number = 0;
 
 	// which group (drawing stripe) is hovered at the moment
 	hoveredGroup: string = null;
 	openedGroupList: boolean = true;
-	
+
 
 	vizScale: number = 1;	
-	
+
 	// store siubscription to unsubscribe on destroy
 	paramsSubscription:Subscription;
-	
-	constructor(private router: Router, private route: ActivatedRoute, private dataService: DataService, private _toastService: ToastService){
+
+	constructor(private router: Router, private route: ActivatedRoute, private dataService: DataService, private codelistService:CodelistService, private _toastService: ToastService){
 		this.groups = [
 			{id: "1", title: "Daňové příjmy"},
 			{id: "2", title: "Nedaňové příjmy"},
 			{id: "3", title: "Kapitálové příjmy"},
 			{id: "4", title: "Přijaté transfery"}
 		];	
-		
+
 		this.groups = [
 			{id:  "11", title: "Daně z příjmů, zisku a kapitálových výnosů", eventsColumnHeader: "Příjem do rozpočtu obce"},
 			{id:  "12", title: "Daně ze zboží a služeb v tuzemsku", eventsColumnHeader: "Příjem do rozpočtu obce"},
@@ -88,34 +89,33 @@ export class IncomeVizComponent{
 
 
 	ngOnInit(){
-		
+
 		this.paramsSubscription = this.route.params.subscribe((params: Params) => {		
-			
+
 			let newState = {
 				group: this.groupIndex[params["skupina"]] ? params["skupina"] : null,
 				year: Number(params["rok"]),
 				event: params["akce"]
 			};
-			
+
 			let oldState = this.state;
-			
+
 			this.openedGroupList = !!newState.group;
-			
+
 			if(newState.year !== oldState.year && newState.year){
-				this.loadBudget(this.profile._id,newState.year);
-				this.loadEvents(this.profile._id,newState.year);
+				this.loadData(this.profile._id,newState.year);
 			}
-			
+
 			if(newState.event !== oldState.event && this.modalLoaded){
 				if(newState.event) this.eventReceiptsModal.show();
 				else this.eventReceiptsModal.hide();
 			}
-			
+
 			this.state = newState;
-			
+
 		});
-		
-  }
+
+	}
 
 	ngAfterViewInit(){
 		this.modalLoaded = true;
@@ -140,54 +140,45 @@ export class IncomeVizComponent{
 	}
 
 	updateState(setParams){
-		
+
 		var replaceUrl = !this.state.year; // if no year was set replace history so we can go back
-		
+
 		let params = {
 			rok: "year" in setParams ? setParams.year : this.state.year,
 			skupina: "group" in setParams ? setParams.group : this.state.group,
 			akce: "event" in setParams ? setParams.event : this.state.event
 		};
-		
+
 		Object.keys(params).forEach(key => { if(!params[key]) delete params[key];});
-		
+
 		this.router.navigate(["./",params],{relativeTo:this.route, replaceUrl:replaceUrl});
 	}
-		
+
 
 	/**
 		* method to handle left/right arrows to switch the selected group
 		*/
 	hotkeys(event){
-		
+
 		// we need to get array of groups so we can get next/prev group
 		var groupIds = Object.keys(this.groupIndex);
-		
+
 		// index of current group. returns -1 in case no group selected, which is no problem for us
 		var i = groupIds.indexOf(this.state.group);
 
 		//LEFT
 		if(event.keyCode == 37) this.selectGroup(groupIds[i - 1 >= 0 ? i - 1 : groupIds.length - 1]);
-		
+
 		//RIGHT
 		if(event.keyCode == 39) this.selectGroup(groupIds[i + 1 <= groupIds.length - 1 ? i + 1 : 0]);
-  }
+	}
 
-	
-
-	 // numbers are parsed from CSV as text
+	// numbers are parsed from CSV as text
 	string2number(string){
 		if(string.charAt(string.length - 1) === "-") string = "-" + string.substring(0,string.length - 1); //sometimes minus is at the end, put it to first character
 		string.replace(",","."); // function Number accepts only dot as decimal point
 		return Number(string);																									
 	}
-
-	getEventName(eventId){
-		if(this.eventIndex[eventId]) return this.eventIndex[eventId].name;
-		if(eventId) return "Nepojmenovaná investiční akce";
-		return "Ostatní";
-	}
-
 
 	getDonutChartData(item){
 		return {
@@ -198,20 +189,14 @@ export class IncomeVizComponent{
 
 	/* PROCESS DATA */
 
-	loadEvents(profileId,year){
-		// get event names
-		this.dataService.getProfileEvents(profileId,{year:year})
-			.then(events => {
-				this.events = events;
-				this.eventIndex = {};
-				this.events.forEach(event => this.eventIndex[event._id] = event);
-			});
-	}
+	loadData(profileId,year){		
+		var queue = [];
 
-	loadBudget(profileId,year){
-		// we get a Promise
-		this.dataService.getProfileBudget(profileId,year)
-			.then((budget) => this.setData(budget))
+		queue.push(this.codelistService.getCodelist("item-names",new Date(year,0,1)));
+		queue.push(this.dataService.getProfileBudget(profileId,year));
+		queue.push(this.dataService.getProfileEvents(profileId,{year:year}));
+		Promise.all(queue)
+			.then(values => this.setData(values[0],values[1],values[2])) // values[0]=item names, values[1]=budget, values[2]=events
 			.catch((err) => {
 			switch(err.status){
 				case 404:
@@ -228,30 +213,64 @@ export class IncomeVizComponent{
 		});
 	}
 
-  setData(budget){
-		
+	setData(itemNames, budget, events){
+		// set paragraphNames;
+		this.itemNames = itemNames;
+		console.log(budget);
+		// create event index
+		let eventIndex = {};
+		events.forEach(event => eventIndex[event._id] = event);
+
 		// set group values at once
-    this.groups.forEach(group => {
+		this.groups.forEach(group => {
 			group.amount = 0;
 			group.budgetAmount = 0;
 			group.items = [];
-    });
-		
+		});
+
 		budget.items.forEach(item => {
 
 			let groupId = item.id.substring(0,2);
-			
-			if(this.groupIndex[groupId]){
-				this.groupIndex[groupId].budgetAmount += item.budgetIncomeAmount;
-				this.groupIndex[groupId].amount += item.incomeAmount;
-				this.groupIndex[groupId].items.push(item);
+			let group = this.groupIndex[groupId];
+
+			// this shouldnt happen, but it might
+			if(!group) return;
+
+			let eventAmount = 0;
+			let eventBudgetAmount = 0;
+
+			item.events.forEach(event => {
+				// assign name from event index
+				event.name = eventIndex[event.event] ? eventIndex[event.event].name : "Nepojmenovaný příjem";
+				// sum events to get the Other value
+				eventAmount += event.incomeAmount;
+				eventBudgetAmount += event.budgetIncomeAmount;
+			});
+
+			// sort events in paragraphs
+			item.events.sort((a,b) => a.name.localeCompare(b.name));
+
+			// add Other if necessary
+			if(item.incomeAmount !== eventAmount || item.budgetIncomeAmount !== eventBudgetAmount){
+				item.events.push({
+					name: "Ostatní",
+					incomeAmount: item.incomeAmount - eventAmount,
+					budgetIncomeAmount: item.budgetIncomeAmount - eventBudgetAmount
+				});
 			}
+
+			group.budgetAmount += item.budgetIncomeAmount;
+			group.amount += item.incomeAmount;
+
+			// add the paragraph to the paragraph list
+			group.items.push(item);
+
 		});
-		
+
+		this.maxAmount = 0;
 		this.groups.forEach(group => {
 			this.maxAmount = Math.max(this.maxAmount,group.budgetAmount,group.amount);
 		});
-		
-  }
+	}
 
 }
